c化图模式bug修复 1. api_event_db_dumper.cpp: 适配model层的api 2. cann_analysis_gear.py: 将node一对多task的方案修改为全量task都记录（之前只要第一条） 3. cann_trace_db_dumper.cpp: requestId字段类型修正为带符号整数 4. event_grouper.cpp: 避免冗余的api异常日志 5. fake_trace_generator.h: api支持设置model_id和request_id 6. tree_analyzer.cpp: 更严格的计算算子判断逻辑 7. tree_builder.cpp: 修复bug，该bug导致建树错误 如model[0,200],node[1,100],task[150,160]，task不应该连到root而是model
# message auto-generated for no-merge-commit merge:
merge new_new into master

c化bug修复

Created-by: msprofDeveloper

Match-id-d2c6fd4f02a2667e3665b520ab0633c47d25f9db

+++ analysis/csrc/viewer/database/drafts/cann_trace_db_dumper.cpp
@@ -31,7 +31,7 @@ using HCCLOpsDumpData = std::vector<std::tuple<uint32_t, uint64_t, int32_t, uint
         std::string, uint64_t, uint64_t, std::string, int64_t>>;
 
 using HostTasksDumpData = std::vector<std::tuple<uint32_t,
-        uint32_t, uint32_t, uint32_t, std::string, uint32_t, std::string, uint32_t, std::string, int64_t>>;
+        int64_t, uint32_t, uint32_t, std::string, uint32_t, std::string, uint32_t, std::string, int64_t>>;
 
 using HcclTasksDumpData = std::vector<std::tuple<uint32_t, int32_t, std::string, std::string, uint32_t, std::string,
         double, uint32_t, uint32_t, uint32_t, uint32_t, uint32_t, uint32_t, uint32_t, uint32_t,
